requirements.txt
# Core dependencies
ultralytics>=8.3.0  # Latest YOLO version for faster training
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Dataset handling
SoccerNet>=2.0.0
huggingface-hub>=0.16.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization and monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.8.0

# Optional: For faster data loading
nvidia-dali-cuda120>=1.28.0; sys_platform != "darwin"

